# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BackBet API",
    "version": "1.0.0",
    "description": "API de Apostas - Backend",
    "contact": {
      "name": "BackBet Team",
      "url": "https://github.com/dioneyfroes-coder/backBet"
    },
    "license": {
      "name": "ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Desenvolvimento"
    },
    {
      "url": "https://api.backbet.com/api",
      "description": "Produção"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token ou User ID em desenvolvimento"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "cadaeb28-c7f7-425b-91f7-73a27141ae49"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "username": {
            "type": "string",
            "example": "joaosilva"
          },
          "firstName": {
            "type": "string",
            "example": "João"
          },
          "lastName": {
            "type": "string",
            "example": "Silva"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING_VERIFICATION",
              "ACTIVE",
              "SUSPENDED"
            ],
            "example": "ACTIVE"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-11-14T23:20:42.778Z"
          }
        },
        "required": [
          "id",
          "email",
          "username",
          "status",
          "createdAt"
        ]
      },
...
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "minLength": 8,
            "example": "SecurePass@123"
          },
          "firstName": {
            "type": "string",
            "minLength": 2,
            "example": "João"
          },
          "lastName": {
            "type": "string",
            "minLength": 2,
            "example": "Silva"
          },
          "username": {
            "type": "string",
            "minLength": 3,
            "pattern": "^[a-zA-Z0-9_]+$",
            "example": "joaosilva"
          }
        },
        "required": [
          "email",
          "password",
          "firstName",
          "lastName",
          "username"
        ]
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "accessToken": {
                "type": "string",
                "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
              },
              "refreshToken": {
                "type": "string",
                "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
              },
              "user": {
                "$ref": "#/components/schemas/User"
              }
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "timestamp": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        },
        "example": {
          "success": true,
          "data": {
            "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.example",
            "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.example",
            "user": {
              "id": "cadaeb28-c7f7-425b-91f7-73a27141ae49",
              "email": "user@example.com",
              "username": "joaosilva",
              "firstName": "João",
              "lastName": "Silva",
              "status": "PENDING_VERIFICATION",
              "createdAt": "2025-11-14T23:20:42.778Z"
            }
          },
          "meta": {
            "timestamp": "2025-12-04T15:40:17.562Z"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "BAD_REQUEST"
              },
              "message": {
                "type": "string",
                "example": "Dados inválidos"
              },
              "details": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/ValidationError"
                  },
                  {
                    "type": "object"
                  }
                ]
              }
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "timestamp": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        },
        "example": {
          "success": false,
          "error": {
            "code": "BAD_REQUEST",
            "message": "Dados inválidos",
            "details": {
              "username": "Invalid input: expected string"
            }
          },
          "meta": {
            "timestamp": "2025-12-04T15:40:17.562Z"
          }
        }
      },
      "RegisterResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "example": "Usuário registrado com sucesso"
              },
              "registrationRequestId": {
                "type": "string",
                "format": "uuid",
                "example": "cadaeb28-c7f7-425b-91f7-73a27141ae49"
              },
              "status": {
                "type": "string",
                "enum": [
                  "PENDING_VERIFICATION",
                  "ACTIVE",
                  "SUSPENDED"
                ],
                "example": "ACTIVE"
              },
              "isActive": {
                "type": "boolean",
                "example": true
              },
              "accessToken": {
                "type": [
                  "string",
                  "null"
                ],
                "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
              },
              "refreshToken": {
                "type": [
                  "string",
                  "null"
                ],
                "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
              },
              "sessionId": {
                "type": [
                  "string",
                  "null"
                ],
                "example": "session-uuid"
              },
              "user": {
                "$ref": "#/components/schemas/User"
              },
              "wallet": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "example": "cadaeb28-c7f7-425b-91f7-73a27141ae49"
                  },
                  "userId": {
                    "type": "string",
                    "example": "cadaeb28-c7f7-425b-91f7-73a27141ae49"
                  },
                  "balance": {
                    "type": "number",
                    "example": 0
                  },
                  "lockedBalance": {
                    "type": "number",
                    "example": 0
                  },
                  "currency": {
                    "type": "string",
                    "example": "BRL"
                  }
                }
              }
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "timestamp": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        },
        "example": {
          "success": true,
          "data": {
            "message": "Usuário registrado com sucesso",
            "registrationRequestId": "cadaeb28-c7f7-425b-91f7-73a27141ae49",
            "status": "ACTIVE",
            "isActive": true,
            "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "sessionId": "session-uuid",
            "user": {
              "id": "cadaeb28-c7f7-425b-91f7-73a27141ae49",
              "email": "user@example.com",
              "username": "joaosilva",
              "status": "PENDING_VERIFICATION",
              "createdAt": "2025-11-14T23:20:42.778Z"
            },
            "wallet": {
              "id": "cadaeb28-c7f7-425b-91f7-73a27141ae49",
              "userId": "cadaeb28-c7f7-425b-91f7-73a27141ae49",
              "balance": 0,
              "lockedBalance": 0,
              "currency": "BRL"
            }
          },
          "meta": {
            "timestamp": "2025-11-14T23:20:42.778Z"
          }
        }
      },
      "MeResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/User"
          },
          "meta": {
            "type": "object",
            "properties": {
              "timestamp": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        },
        "example": {
          "success": true,
          "data": {
            "id": "cadaeb28-c7f7-425b-91f7-73a27141ae49",
            "email": "user@example.com",
            "username": "joaosilva",
            "firstName": "João",
            "lastName": "Silva",
            "status": "ACTIVE",
            "createdAt": "2025-11-14T23:20:42.778Z"
          },
          "meta": {
            "timestamp": "2025-11-14T23:21:06.343Z"
          }
        }
      },
      "LogoutResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "example": "Logout realizado com sucesso"
              }
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "timestamp": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        },
        "example": {
          "success": true,
          "data": {
            "message": "Logout realizado com sucesso"
          },
          "meta": {
            "timestamp": "2025-11-14T23:22:00.000Z"
          }
        }
      },
      "ConflictError": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "CONFLICT"
              },
              "message": {
                "type": "string",
                "example": "Email já cadastrado"
              }
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "timestamp": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        },
        "example": {
          "success": false,
          "error": {
            "code": "CONFLICT",
            "message": "Email já cadastrado"
          },
          "meta": {
            "timestamp": "2025-11-14T23:20:00.000Z"
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "description": "Detalhes de validação por campo",
        "additionalProperties": {
          "type": "string",
          "example": "Campo obrigatório"
        },
        "example": {
          "email": "Formato de email inválido",
          "username": "Tamanho mínimo: 3"
        }
      },
...
      "UnauthorizedError": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "UNAUTHORIZED"
              },
              "message": {
                "type": "string",
                "example": "Autenticação requerida"
              }
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "timestamp": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        },
        "example": {
          "success": false,
          "error": {
            "code": "UNAUTHORIZED",
            "message": "Autenticação requerida"
          },
          "meta": {
            "timestamp": "2025-11-14T23:21:00.000Z"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Registra novo usuário e cria carteira inicial",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário registrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Email já cadastrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Autentica usuário via credenciais e retorna tokens JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "example": "Senha"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token retornado",
            "content": {
              "application/json": {
                "$ref": "#/components/schemas/AuthResponse"
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas",
            "content": {
              "application/json": {
                "$ref": "#/components/schemas/ErrorResponse"
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Renova access token usando refresh token (placeholder)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token renovado",
            "content": {
              "application/json": {
                "$ref": "#/components/schemas/AuthResponse"
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "$ref": "#/components/schemas/ErrorResponse"
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Retorna dados do usuário autenticado",
        "responses": {
          "200": {
            "description": "Dados do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Faz logout e invalida sessão no cliente",
        "responses": {
          "200": {
            "description": "Logout realizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          }
        }
      }
    },
...
  }
}